projects/project5_1/requirements.txt
crewai
crewai-tools
langchain_openai
python-dotenv
pyyaml
